+++ pysm3/tests/test_interpolating.py
@@ -13,7 +13,7 @@ def test_interpolating(tmp_path):
     hp.write_map(tmp_path / "20.fits", np.ones(shape, dtype=np.float32))
 
     interp = InterpolatingComponent(
-        tmp_path, "uK_RJ", nside, interpolation_kind="linear", has_polarization=True
+        tmp_path, "uK_RJ", nside, interpolation_kind="linear"
     )
 
     interpolated_map = interp.get_emission(15 * u.GHz)